fix: 인벤토리 가격 출력 제거 및 HP 타입 불일치 경고 수정
- displayItems()에서 가격 출력 제거 (가격은 상점에서만 표시)
- Player::setMaxHP(), setCurHP() 인자 타입 float → int 변경
(Inventory에서 std::min 사용 시 발생하던 C4244 경고 해결)

+++ Player/Player/Player.cpp
@@ -148,12 +148,12 @@ void Player::setExp(int exp)
 	this->exp = exp;
 }
 
-void Player::setMaxHP(float maxHp)
+void Player::setMaxHP(int maxHp)
 {
 	this->maxHp = maxHp;
 }
 
-void Player::setCurHP(float curHp)
+void Player::setCurHP(int curHp)
 {
 	this->curHp = curHp;
 }
